enable wheat seed
